+++ src/main/java/randommcsomethin/fallingleaves/util/TextureCache.java
@@ -5,7 +5,7 @@ import net.minecraft.util.Identifier;
 import java.util.HashMap;
 
 public class TextureCache {
-    public static class Data {
+    public static final class Data {
         public final int color;
         public final String resourcePack;
 
@@ -18,4 +18,5 @@ public class TextureCache {
     public static final HashMap<Identifier, Data> INST = new HashMap<>();
 
     private TextureCache() {}
+
 }